requirements.txt
wheel
Django==5.0.6
djangorestframework==3.15.1
Markdown==3.6
django-filter==24.2
